% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualtabulation.R
\name{air_annual}
\alias{air_annual}
\title{Get annual tabulations}
\usage{
air_annual(component, year)
}
\arguments{
\item{component}{ID of component (see \link{air_components})}

\item{year}{A 4-digit year}
}
\description{
Returns the annual balances for a component by station
}
\examples{
\dontrun{
air_annual(5, 2020)
}
}
