% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{air_data_availability}
\alias{air_data_availability}
\title{Get data availability}
\usage{
air_data_availability(
  use = c("annualbalance", "airquality", "measure"),
  date_from = NULL,
  time_from = NULL,
  date_to = NULL,
  time_to = NULL
)
}
\arguments{
\item{use}{Should data for "annualbalance", "airquality", or "measure" be returned?}

\item{date_from}{A date ("YYYY-MM-DD", required only for use = "airquality")}

\item{time_from}{An hour (1..24)}

\item{date_to}{A date ("YYYY-MM-DD", required only for use = "airquality")}

\item{time_to}{An hour (1..24)}
}
\description{
Returns a list of available dates/years
}
