% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceedances.R, R/metadata.R
\name{air_exceedances}
\alias{air_exceedances}
\alias{air_exceedances_types}
\title{Get exceedances data}
\usage{
air_exceedances(component, year)

air_exceedances_types()
}
\arguments{
\item{component}{ID of component (see \link{air_components})}

\item{year}{A 4-digit year}
}
\description{
\code{air_exceedances()} returns exceedances for a component by station and month.

\code{air_exceedances_types()} returns a list of all exceedances types.
}
\examples{
\dontrun{
air_exceedances(5, 2020)
}
}
