% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\name{air_measures}
\alias{air_measures}
\alias{air_measures_date_limits}
\title{Get all measurements}
\usage{
air_measures(
  date_from,
  time_from,
  date_to,
  time_to,
  station = NULL,
  component = NULL,
  scope = NULL
)

air_measures_date_limits()
}
\arguments{
\item{date_from}{A date ("YYYY-MM-DD", required only for use = "airquality")}

\item{time_from}{An hour (1..24)}

\item{date_to}{A date ("YYYY-MM-DD", required only for use = "airquality")}

\item{time_to}{An hour (1..24)}

\item{station}{ID of the \link[=air_stations]{station} to get data from. If \code{NULL},
include all available stations.}

\item{component}{ID of \link[=air_components]{component}. If NULL, include all available components.}

\item{scope}{ID of \link[=air_scopes]{scope}. If NULL, include all available scopes.}
}
\value{
A named list where the names are station IDs and each element is a
data frame for the corresponding station
}
\description{
\code{air_measures()} returns measures for the provided parameters.

\code{air_measures_date_limits()} returns the date limits of measurements by
\link[=air_scopes]{scope}, \link[=air_components]{component}, and \link[=air_stations]{station}.
}
\examples{
\dontrun{
air_measures("2020-01-01", 9, "2020-01-01", 11)
}
}
