% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{air_components}
\alias{air_components}
\alias{air_networks}
\alias{air_scopes}
\alias{air_thresholds}
\alias{air_data_availability}
\title{Get metadata}
\usage{
air_components()

air_networks()

air_scopes()

air_thresholds(use, component = NULL, scope = NULL)

air_data_availability(
  use = c("annualbalance", "airquality", "measure"),
  date_from = NULL,
  time_from = NULL,
  date_to = NULL,
  time_to = NULL
)
}
\arguments{
\item{use}{Should data for "annualbalance", "airquality", or "measure" be returned?}

\item{component}{ID of component (see \link{air_components})}

\item{scope}{ID of scope (see \link{air_scopes})}

\item{date_from}{A date ("YYYY-MM-DD", required only for use = "airquality")}

\item{time_from}{An hour (1..24)}

\item{date_to}{A date ("YYYY-MM-DD", required only for use = "airquality")}

\item{time_to}{An hour (1..24)}
}
\description{
\code{air_components()} returns a list of all air components measured.
\code{air_networks()} returns a list of all air measuring station networks.
\code{air_scopes()} returns a list of all measurement scopes.
\code{air_thresholds()} returns a list of all thresholds.
\code{air_data_availability()} returns a list of available dates/years.
}
