% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{air_stations}
\alias{air_stations}
\alias{air_station_settings}
\alias{air_station_types}
\title{Get stations}
\usage{
air_stations(
  use = c("annualbalance", "airquality", "measure", "transgression"),
  date_from = NULL,
  time_from = NULL,
  date_to = NULL,
  time_to = NULL,
  return_sf = FALSE
)

air_station_settings()

air_station_types()
}
\arguments{
\item{use}{Should data for "annualbalance", "airquality", "measure", or
"transgression" be returned?}

\item{date_from, date_to}{A date ("YYYY-MM-DD", required only for use = "airquality")}

\item{time_from, time_to}{An hour (1..24)}
}
\description{
\code{air_stations()} returns a list of all air measuring stations for a certain use.

\code{air_station_settings()} returns a list of all station settings.

\code{air_station_types()} returns a list of all station types.
}
