% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{air_thresholds}
\alias{air_thresholds}
\title{Get all thresholds}
\usage{
air_thresholds(use, component = NULL, scope = NULL)
}
\arguments{
\item{use}{Should thresholds for "airquality" or "measure" be returned?}

\item{component}{ID of component (see \code{\link[=air_components]{air_components()}})}

\item{scope}{ID of scope (see \link{air_scopes})}
}
\description{
Returns a list of all station types
}
