% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airquality.R
\name{air_quality}
\alias{air_quality}
\alias{air_quality_date_limits}
\title{Get air quality data}
\usage{
air_quality(date_from, time_from, date_to, time_to, station = NULL)

air_quality_date_limits()
}
\arguments{
\item{date_from}{A date ("YYYY-MM-DD", required only for use = "airquality")}

\item{time_from}{An hour (1..24)}

\item{date_to}{A date ("YYYY-MM-DD", required only for use = "airquality")}

\item{time_to}{An hour (1..24)}

\item{station}{ID of the \link[=air_stations]{station} to get data from. If \code{NULL},
include all available stations.}
}
\value{
A named list where the names are station IDs and each element is a
\href{https://tidyr.tidyverse.org/articles/nest.html}{nested data frame} for the
corresponding station with 5 columns:
\itemize{
\item start_dttm, end_dttm: Time of start and end of measuring in CET
\item index: Airquality index for all components
\item is_incomplete: Flag if data is incomplete (not all components available)
\item data: List column of data frames with the measured components
\itemize{
\item component: Component ID
\item value: Measured value for this component
\item index: Airquality index of this component
\item pct_in_index_thresholds: Decimal representation of percent in index
thresholds. 0.x is x percent between index 0 and 1, 1.x is x percent
between index 1 and 2 etc.
}
}
}
\description{
\code{air_quality()} returns air quality data for the provided parameters.

\code{air_quality_date_limits()} returns the date limits of air quality data by
\link[=air_stations]{station}.
}
\examples{
\dontrun{
air_quality("2020-01-01", 9, "2020-01-01", 11, station = 7)
}
}
